requirements.txt
requests
pypdf
python-docx
openai
openpyxl
pandas
xlswriter

